double fields

+++ src/test/java/org/elasticsearch/opentender/nativescript/script/ScriptedMetricAggsScriptTests.java
@@ -44,8 +44,8 @@ public class ScriptedMetricAggsScriptTests extends AbstractSearchScriptTestCase 
         String stockMapping = XContentFactory.jsonBuilder().startObject().startObject("stock")
                 .startObject("properties")
                 .startObject("type").field("type", "string").field("index", "not_analyzed").endObject()
-                .startObject("value1").field("type", "long").endObject()
-                .startObject("value2").field("type", "long").endObject()
+                .startObject("value1").field("type", "double").endObject()
+                .startObject("value2").field("type", "double").endObject()
                 .endObject().endObject().endObject()
                 .string();
 
@@ -97,8 +97,8 @@ public class ScriptedMetricAggsScriptTests extends AbstractSearchScriptTestCase 
         String stockMapping = XContentFactory.jsonBuilder().startObject().startObject("stock")
                 .startObject("properties")
                 .startObject("type").field("type", "string").field("index", "not_analyzed").endObject()
-                .startObject("value1").field("type", "long").endObject()
-                .startObject("value2").field("type", "long").endObject()
+                .startObject("value1").field("type", "double").endObject()
+                .startObject("value2").field("type", "double").endObject()
                 .endObject().endObject().endObject()
                 .string();
 
@@ -154,8 +154,8 @@ public class ScriptedMetricAggsScriptTests extends AbstractSearchScriptTestCase 
         String stockMapping = XContentFactory.jsonBuilder().startObject().startObject("stock")
                 .startObject("properties")
                 .startObject("type").field("type", "string").field("index", "not_analyzed").endObject()
-                .startObject("value1").field("type", "long").endObject()
-                .startObject("value2").field("type", "long").endObject()
+                .startObject("value1").field("type", "double").endObject()
+                .startObject("value2").field("type", "double").endObject()
                 .endObject().endObject().endObject()
                 .string();
 
@@ -211,8 +211,8 @@ public class ScriptedMetricAggsScriptTests extends AbstractSearchScriptTestCase 
         String stockMapping = XContentFactory.jsonBuilder().startObject().startObject("stock")
                 .startObject("properties")
                 .startObject("type").field("type", "string").field("index", "not_analyzed").endObject()
-                .startObject("value1").field("type", "long").endObject()
-                .startObject("value2").field("type", "long").endObject()
+                .startObject("value1").field("type", "double").endObject()
+                .startObject("value2").field("type", "double").endObject()
                 .endObject().endObject().endObject()
                 .string();
